use the HEAD version of doctest-discover
Unlike doctest-discover-0.1.0.7, that version finds the files in
subfolders as well. (I'm about to add a new file in a subfolder)

I can't specify that version in the cabal file, which means travis won't
run the tests in the subfolder. I should switch back to the stack
configuration and figure out why it didn't work... But I think I'll just
wait for the next doctest-dicover release instead.

[ci skip]